reverted drawBitmap back to old library use until I can fix it

+++ master/Adafruit_GFX/Adafruit_GFX.h
@@ -72,7 +72,7 @@ public:
     void fillRoundRect(int16_t x0, int16_t y0, int16_t w, int16_t h, int16_t radius, uint16_t color);
 
     void drawBitmap(int16_t x, int16_t y,
-                    const uint16_t *bitmap, int16_t w, int16_t h,
+                    const uint8_t *bitmap, int16_t w, int16_t h,
                     uint16_t color);
     void drawChar(int16_t x, int16_t y, unsigned char c,
                   uint16_t color, uint16_t bg, uint8_t size);
